% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{adjacencyMatrix}
\alias{adjacencyMatrix}
\title{adjacencyMatrix data set}
\description{
This RData file contains an adjacency matrix that is intended to be used as input for spectral clustering using the spectralClustering function.
}
\details{
Adjacency Matrix for Spectral Clustering

This RData file contains an adjacency matrix that is intended to be used as input
for spectral clustering using the spectralClustering function. Spectral clustering
is a graph-based clustering technique often applied to such adjacency matrices.

The adjacency matrix represents the connectivity of a graph, where each row and
column corresponds to a node, and the matrix entries indicate the edges between
nodes (1 for connected, 0 for not connected).
}
\examples{
\dontrun{
data(adjacency_matrix)
numClusters <- 2
result <- spectralClustering(adjacencyMatrix, numClusters)
cat("Community Assignments:", result, "\n")
}
}
