% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of R functions (\code{norm_clus}) and Cpp functions (\code{norm_clus_C}).
}
\examples{
\dontrun{
data(data_mix2)
tm1 <- microbenchmark::microbenchmark(
  rnR = norm_clus(data_mix2,10,iters=FALSE),
  rnC = norm_clusC(data_mix2,10,iters=FALSE))
print(summary(tm1)[,c(1,3,5,6)])
}
}
