% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{data_mix2}
\alias{data_mix2}
\title{Binary mixed normal data set}
\description{
Is an example of a data set with a length of 100 and a distribution of 0.6N(0,1)+0.4N(8,4).
}
\examples{
\dontrun{
data(data_mix2)
clu<- norm_clus(data_mix2,2)
clu$label
}
}
