% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_clusC}
\alias{norm_clusC}
\title{Classify the mixed normal data using Rcpp}
\usage{
norm_clusC(data, K, iters = TRUE)
}
\arguments{
\item{data}{Data set for clustering, length \code{n}}

\item{K}{EM Indicates the maximum number of iterations of the algorithm}

\item{iters}{Whether to output EM number of iterations, the default value is T}
}
\value{
A list consisting of the following sections:
\describe{
  \item{\code{label}}{Data of length \code{n} consisting of 0 or 1 or 2, used to mark the classification category}
  \item{\code{param}}{Estimates of the mean and standard deviation of the corresponding two or three normal distributions, as well as estimates of the mixed proportions}
}
}
\description{
The EM algorithm is used to classify data from binary or ternary positive mixing models Rcpp
}
\examples{
\dontrun{
data(data_mix2)
norm_clusC(data_mix2,10)
}

}
