\name{SA23204182-package}
\alias{SA23204182-package}
\alias{SA23204182}
\docType{package}
\title{
  Final project
}
\description{
  Implement some simple algorithms.
}
\author{
Zhang Yan

Maintainer: Zhang Yan <zy19050972@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
